requirements.txt
pandas
python-dotenv
mysql-connector-python
boto3
zipfile36